Method ignores exceptional return value

+++ src/main/java/org/mule/module/mongo/tools/BsonDumpWriter.java
@@ -42,6 +42,10 @@ public class BsonDumpWriter extends DumpWriter
         FileOutputStream outputStream = null;
         File outputFile = new File(getFilePath(collection));
         outputFile.getParentFile().mkdirs();
+        if(!outputFile.exists() && !outputFile.mkdirs())
+        {
+            throw new IllegalStateException("Couldn't create dir: " + outputFile);
+        }
         try
         {
             outputStream = new FileOutputStream(outputFile, true);
